+++ src/index.tsx
@@ -22,6 +22,7 @@ import Loader from 'loader'
 import Menu from 'menu'
 import Message from 'message'
 import Modal from 'modal'
+import PhoneField from 'phoneField'
 import Profile from 'profile'
 import Proximity from 'proximity'
 import PushInfo from 'pushInfo'
@@ -71,6 +72,7 @@ export {
   Menu,
   Message,
   Modal,
+  PhoneField,
   Profile,
   Proximity,
   PushInfo,