+++ src/typings/vendors.d.ts
@@ -15,3 +15,9 @@ declare module 'exenv' {
   const canUseEventListeners: () => boolean
   const canUseDOM: () => boolean
 }
+declare module 'country-telephone-data' {
+  const countryTelData: countryTelData
+  const allCountries: allCountriesPhoneData[]
+  const allCountryCodes: allCountryCodes
+  const iso2Lookup: iso2Lookup
+}